refactor Sampler into policies
